Succulent plants/Linebot分類+疾病偵測模型/辨識/測試用/requirements.txt
line-bot-sdk
flask
pillow
tensorflow
uwsgi
torch
typing 
torchvision